backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
langdetect==1.0.9
scikit-learn==1.5.2
openai>=1.40.0
python-dateutil>=2.9.0
opendataloader-pdf>=2.2.0
